requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
websockets==12.0
python-dotenv==1.0.1
faster-whisper==1.0.1
edge-tts==7.0.4
groq==0.4.2
google-generativeai==0.4.1
reportlab==4.1.0
pdfplumber==0.11.0
PyPDF2==3.0.1
